Add files via upload
To run the spider and store the scraped data in a CSV file, you can use the scrapy runspider command:

Copy code
scrapy runspider ebay_price.py -o prices.csv
This will run the spider and store the scraped data in a CSV file named prices.csv. You can also use the -a option to pass arguments to the spider, such as the rate limit or the number of pages to scrape.

Copy code
scrapy runspider scrapy scraper.py  -a rate_limit=1 -a max_pages=10 -o prices.csv
This will run the spider with a rate limit of 1 request per second and a maximum of 10 pages, and store the scraped data in a CSV file named prices.csv.