Update scrape ebay.py
Had to update to save item item name, price and the date and time it saved at in order to  make it useful for gathering a price history. I'm now beginning to think it might be easier to turn this into a SQL database and use MariaDB to manage it in the future but... For now.