Add files via upload
Well the structure of everything changed, I changed it to take arguments from the user  via command line rather than  editting the file to receive the file. It ignores robots.txt,  It also now uses selenium stealth and  rotates a Good amount of user ages to gather this. Here's for being able to compile a free price history for you folk.